+++ lib/ig_markets/streaming/account_state.rb
@@ -171,8 +171,12 @@ module IGMarkets
         @working_orders.each do |working_order|
           next unless working_order.deal_id == working_order_update.deal_id
 
-          working_order.limit_distance = working_order_update.limit_distance
-          working_order.stop_distance = working_order_update.stop_distance
+          working_order.currency_code = working_order_update.currency
+          working_order.order_level = working_order_update.level
+
+          %i(good_till_date limit_distance order_type stop_distance time_in_force).each do |attribute|
+            working_order.send "#{attribute}=", working_order_update.send(attribute)
+          end
         end
       end